chrome/test/data/xr/e2e_test_files/html/test_webxr_does_not_return_device.html

<!doctype html>
<!--
Tests that WebXR does not report immersive-vr support when OpenVR is disabled.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <canvas id="webgl-canvas"></canvas>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script>
      navigator.xr.isSessionSupported('immersive-vr').then((supported) => {
        assert_false(supported, "WebXR indicated support for 'immersive-vr'");
        done();
      });
    </script>
  </body>
</html>
